:truck: Moved types of parsed objects to userscript/*.ts
